+++ Frameworks/OpenMPT/OpenMPT/openmpt123/openmpt123.cpp
@@ -46,6 +46,7 @@ static const char * const license =
 #include <limits>
 #include <locale>
 #include <map>
+#include <memory>
 #include <random>
 #include <set>
 #include <sstream>